src/{enums, parsers}: use Vec<16> for SGR
This bumps the amount of SGR codes that are allowed to 16, based on the
`man 4 console_codes` page. The parser appears to be based off of single
use valid SGR codes, such as setting the color to true color, but
doesn't make use of the fact that SGR codes can be chained one after
another.